[move] subfolders into common directory
