+++ src/editor/common/configuration/editorConfiguration.ts
