+++ src/editor/common/extensions/editorExtension.ts
