+++ src/editor/editorWidget.ts
@@ -16,9 +16,9 @@ import { REGISTRANTS } from "src/code/platform/registrant/common/registrant";
 import { IEditorModel } from "src/editor/common/model";
 import { IEditorView } from "src/editor/common/view";
 import { IEditorViewModel } from "src/editor/common/viewModel";
-import { EditorOptions, EditorOptionsType, IEditorOption, IEditorWidgetOptions } from "src/editor/configuration/editorConfiguration";
-import { IEditorExtension } from "src/editor/contrib/extensions/editorExtension";
-import { IEditorExtensionRegistrant } from "src/editor/contrib/extensions/editorExtensionRegistrant";
+import { EditorOptions, EditorOptionsType, IEditorOption, IEditorWidgetOptions } from "src/editor/common/configuration/editorConfiguration";
+import { IEditorExtension } from "src/editor/common/extensions/editorExtension";
+import { IEditorExtensionRegistrant } from "src/editor/common/extensions/editorExtensionRegistrant";
 import { EditorModel } from "src/editor/model/editorModel";
 import { EditorView } from "src/editor/view/editorView";
 import { EditorViewModel } from "src/editor/viewModel/editorViewModel";
