+++ src/editor/view/editorView.ts
@@ -4,7 +4,7 @@ import { ILogEvent } from "src/base/common/logger";
 import { ProseEditorState } from "src/editor/common/proseMirror";
 import { IEditorView, IEditorViewOptions } from "src/editor/common/view";
 import { EditorRenderType, IEditorViewModel, IRenderEvent } from "src/editor/common/viewModel";
-import { EditorOptionsType } from "src/editor/configuration/editorConfiguration";
+import { EditorOptionsType } from "src/editor/common/configuration/editorConfiguration";
 import { RichtextWindow } from "src/editor/view/viewPart/viewWindow/richtext/richtextWindow";
 import { IViewWindow } from "src/editor/view/viewPart/viewWindow/window";
 
@@ -37,6 +37,7 @@ export class EditorView extends Disposable implements IEditorView {
     constructor(
         container: HTMLElement,
         viewModel: IEditorViewModel,
+        // commandDelegate: ICommandDelegate,
         options: EditorOptionsType,
     ) {
         super();
